+++ ods/lib/safe-env.sh
@@ -175,7 +175,7 @@ load_model_selector_env_from_output() {
         LLAMA_ARG_CACHE_TYPE_V \
         LLAMA_ARG_N_CPU_MOE \
         LLAMA_ARG_NO_CACHE_PROMPT \
-        LLAMA_ARG_CHECKPOINT_EVERY_N_TOKENS \
+        LLAMA_ARG_CHECKPOINT_EVERY_NT \
         LLAMA_ARG_SPEC_TYPE \
         LLAMA_ARG_SPEC_DRAFT_N_MAX \
         LLAMA_ARG_SPEC_DRAFT_TYPE_K \